foreign key added in movies table

+++ vidly/Models/Genre.cs
@@ -7,8 +7,8 @@ using System.Web;
 namespace vidly.Models
 {
     public class Genre
-    {
-        public int Id{get; set;}
+    {    
+        public byte Id{get; set;}
         [Required]
         [StringLength(255)]
         public string GenreName{get; set;}
